feat: Add map helper for direct JS function application.
